Change to local memory allocation for pin arrays
Add destructor to SevSegShift with proper memory cleanup

Fixes issues #7 and #9

Instead of copying just the pointer of the shiftRegisterMapDigits and
shiftRegisterMapSegments arrays passed to begin, we allocate the memory
for them in SegSevShift and copy the values manually.
This removes the implicit need for passing global/static variables, so
begin can be called with local/temporary memory.